requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Pillow==9.2.0
numpy==1.24.2
